+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/tuples/Tuple2.java
@@ -1,23 +1,10 @@
 package io.tech1.framework.domain.tuples;
 
-import lombok.*;
+import lombok.Data;
 
 // Lombok
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-@EqualsAndHashCode
-@ToString
+@Data(staticConstructor = "of")
 public class Tuple2<A, B> {
     private final A a;
     private final B b;
-
-    public static <A, B> Tuple2<A, B> of(
-            A a,
-            B b
-    ) {
-        return new Tuple2<>(
-                a,
-                b
-        );
-    }
 }