+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/tuples/Tuple4.java
@@ -1,29 +1,12 @@
 package io.tech1.framework.domain.tuples;
 
-import lombok.*;
+import lombok.Data;
 
 // Lombok
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-@EqualsAndHashCode
-@ToString
+@Data(staticConstructor = "of")
 public class Tuple4<A, B, C, D> {
     private final A a;
     private final B b;
     private final C c;
     private final D d;
-
-    public static <A, B, C, D> Tuple4<A, B, C, D> of(
-            A a,
-            B b,
-            C c,
-            D d
-    ) {
-        return new Tuple4<>(
-                a,
-                b,
-                c,
-                d
-        );
-    }
 }