+++ tech1-framework-domain/src/test/java/io/tech1/framework/domain/tuples/TupleToggleTest.java
@@ -1,0 +1,41 @@
+package io.tech1.framework.domain.tuples;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import lombok.SneakyThrows;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class TupleToggleTest extends AbstractTupleTest {
+    private static final TupleToggle<String> TUPLE = TupleToggle.enabled("enabled");
+
+    @Override
+    protected String getFileName() {
+        return "tuple-toggle.json";
+    }
+
+    @Test
+    public void serializeTest() {
+        // Act
+        var json = this.writeValueAsString(TUPLE);
+
+        // Assert
+        assertThat(json).isNotNull();
+        assertThat(json).isEqualTo(this.readFile());
+    }
+
+    @SneakyThrows
+    @Test
+    public void deserializeTest() {
+        // Arrange
+        var json = this.readFile();
+        var typeReference = new TypeReference<TupleToggle<String>>() {};
+
+        // Act
+        var tuple = OBJECT_MAPPER.readValue(json, typeReference);
+
+        // Assert
+        assertThat(tuple).isNotNull();
+        assertThat(tuple).isEqualTo(TUPLE);
+    }
+}